+++ angular_ast/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 *   Annotations may now be assigned values.
 
+*   Added support for annotations on `ContainerAst`.
+
 ## 0.5.2
 
 *   The `*` micro-syntax now supports newlines after an identifier.